fix message order in thread, message big count clean up

+++ anfoya-javafx/src/main/java/net/anfoya/javafx/scene/control/IncExcListItem.java
@@ -1,9 +1,9 @@
 package net.anfoya.javafx.scene.control;
 
 import javafx.beans.property.BooleanProperty;
-import javafx.beans.property.IntegerProperty;
+import javafx.beans.property.LongProperty;
 import javafx.beans.property.SimpleBooleanProperty;
-import javafx.beans.property.SimpleIntegerProperty;
+import javafx.beans.property.SimpleLongProperty;
 import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
 
@@ -12,13 +12,13 @@ public class IncExcListItem {
 	private final BooleanProperty disableProperty;
 	private final BooleanProperty includedProperty;
 	private final BooleanProperty excludedProperty;
-	private final IntegerProperty countProperty;
+	private final LongProperty countProperty;
 	private final BooleanProperty focusTraversableProperty;
 
 	public IncExcListItem() {
 		textProperty = new SimpleStringProperty();
 		disableProperty = new SimpleBooleanProperty(true);
-		countProperty = new SimpleIntegerProperty(0);
+		countProperty = new SimpleLongProperty(0);
 		includedProperty = new SimpleBooleanProperty(false);
 		excludedProperty = new SimpleBooleanProperty(false);
 		focusTraversableProperty = new SimpleBooleanProperty(true);
@@ -42,7 +42,7 @@ public class IncExcListItem {
 	public BooleanProperty disableProperty() {
 		return disableProperty;
 	}
-	public IntegerProperty countProperty() {
+	public LongProperty countProperty() {
 		return countProperty;
 	}
 	public BooleanProperty includedProperty() {
